refactor model fixing
