+++ src/ast/macros/CMakeLists.txt
@@ -2,6 +2,7 @@ z3_add_component(macros
   SOURCES
     macro_finder.cpp
     macro_manager.cpp
+    quantifier_macro_info.cpp
     macro_util.cpp
     quasi_macros.cpp
   COMPONENT_DEPENDENCIES
